create custorm exception

+++ src/main/java/com/fico/todo/service/TaskServiceImpl.java
@@ -1,15 +1,16 @@
 package com.fico.todo.service;
 
+import com.fico.todo.exception.OperationNotAllowedException;
+import com.fico.todo.exception.TaskNotFoundException;
 import com.fico.todo.model.Task;
 import com.fico.todo.repository.TaskRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.security.Principal;
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
+import java.util.*;
 
 @Service
 public class TaskServiceImpl implements TaskService{
@@ -27,8 +28,12 @@ public class TaskServiceImpl implements TaskService{
         return taskRepository.save(task);
     }
 
-    public List<Task> findAll(){
-        return taskRepository.findAll();
+    public List<Task> findAll(Principal principal){
+        Set<String> roles = userDetailsService.getRoleSet(principal);
+        if(roles.contains("ADMIN"))
+            return taskRepository.findAll();
+        else
+            throw new OperationNotAllowedException();
     }
 
     public List findByUserId(Optional<Long> userId, Principal principal){
@@ -36,17 +41,26 @@ public class TaskServiceImpl implements TaskService{
         if(principal.getName()==userService.findById(userId).getUsername() || roles.contains("ADMIN")) {
             return taskRepository.findByUserId(userId);
         }
-        return Collections.emptyList();
+        throw new OperationNotAllowedException();
     }
 
     public Optional<Task> findById(Long taskId){
-        return taskRepository.findById(taskId);
+        Optional<Task> task = taskRepository.findById(taskId);
+        if(!task.isPresent()){
+            //return null;
+            throw new TaskNotFoundException();
+        }
+        return task;
     }
 
     public void delete(Long taskId){
-        Task task = taskRepository.getOne(taskId);
-        taskRepository.delete(task);
-
+        if(taskRepository.existsById(taskId)){
+            Task task = taskRepository.getOne(taskId);
+            taskRepository.delete(task);
+        }
+        else {
+            throw new TaskNotFoundException();
+        }
     }
